Core API

A gcdyn.TreeNode is the realization of a gcdyn.AbstractBranchingProcess.

Trees

gcdyn.TreeNode — Type

TreeNode(event, time, type)

A data structure for building realizations of multitype branching processes.

A "tree" is defined as a TreeNode with event = :root and optional TreeNode children of other non-root events. All nodes in the tree (including the root) also have:

a time at which the event occurred

a type attribute whose meaning is left to the user

Note that if a TreeNode has the event of :type_change, the type attribute is the new type along the branch, not the old type.

See also attach!, detach!, TreeTraversal, rand_tree, EVENTS.

gcdyn.rand_tree — Function

rand_tree(model, present_time, init_type [,n]; reject_stubs=true)

Generate n random trees from the given model (optional parameter, default is n=1), each starting at the given initial type.

Note that trees will have root time 0, with time increasing toward the tips. This aligns us with the branching process models, but contradicts the notion of time typically used in phylogenetics.

Can optionally choose not to reject_stubs, equivalent to not conditioning on the root node having at least one sampled descendant. Alternatively, can specify a minimum number of sampled descendants min_leaves that the tree must have. If the number of rejections exceeds max_rejections, an error is thrown.

See also TreeNode, sample_child!, mutate!.

gcdyn.EVENTS — Constant

The possible events that can occur at a node in a TreeNode.

gcdyn.TreeTraversal — Type

TreeTraversal

Abstract supertype for iterators over TreeNodes.

Pre-order and post-order traversals are defined. These are present mainly because the iterators in AbstractTrees were difficult to make type-stable.

See also PostOrderTraversal, PreOrderTraversal.

gcdyn.PostOrderTraversal — Type

PostOrderTraversal(tree)

Iterator for performing a post-order traversal over tree.

See also TreeTraversal.

gcdyn.PreOrderTraversal — Type

PreOrderTraversal(tree)

Iterator for performing a pre-order traversal over tree.

gcdyn.LeafTraversal — Type

LeafTraversal(tree)

Iterator over the leaves of tree.

RecipesBase.plot — Method

plot(tree::TreeNode)

Visualizes the tree.

The optional colorscheme keyword sets the color scheme to use (:linear_kbc_5_95_c73_n256 by default). See ColorSchemes.findcolorscheme for more options.

The optional midpoint keyword sets the type that is equivalent to the median of the colorscheme. Useful for diverging color schemes (:diverging_bkr_55_10_c35_n256 is one recommendation).

The optional reverse_colorscheme keyword reverses the color scheme.

Branching processes

gcdyn.AbstractBranchingProcess — Type

An abstract type for branching processes.

See also SigmoidalBranchingProcess, ConstantBranchingProcess, DiscreteBranchingProcess.

gcdyn.ConstantBranchingProcess — Type

ConstantBranchingProcess(λ, μ, γ, ρ, σ, type_space)
ConstantBranchingProcess(λ, μ, δ, Γ, ρ, σ, type_space)

Constructs a multitype branching process with the given parameters.

A single birth rate λ is provided for all types. The death rate parameter is notated as μ.

Γ is the type change rate matrix, and δ is a scaling parameter for this matrix. Alternatively, a single type change rate γ can be provided for all types, in which case the model will assume uniform transitions between types.

Sampling probabilities are notated as ρ (survival sampling probability) and σ (death sampling probability).

gcdyn.DiscreteBranchingProcess — Type

DiscreteBranchingProcess(λ, μ, γ, ρ, σ, type_space)
DiscreteBranchingProcess(λ, μ, δ, Γ, ρ, σ, type_space)

Constructs a multitype branching process with the given parameters.

A Vector of birth rates λ is provided, with indices corresponding to type_space. The death rate parameter is notated as μ.

Γ is the type change rate matrix, and δ is a scaling parameter for this matrix. Alternatively, a single type change rate γ can be provided for all types, in which case the model will assume uniform transitions between types.

Sampling probabilities are notated as ρ (survival sampling probability) and σ (death sampling probability).

gcdyn.SigmoidalBranchingProcess — Type

SigmoidalBranchingProcess(λ_xscale, λ_xshift, λ_yscale, λ_yshift, μ, γ, ρ, σ, type_space)
SigmoidalBranchingProcess(λ_xscale, λ_xshift, λ_yscale, λ_yshift, μ, δ, Γ, ρ, σ, type_space)

Constructs a multitype branching process with the given parameters.

A parameterized sigmoid curve (with λ_xscale, λ_xshift, λ_yscale, and λ_yshift) maps types to birth rates. The death rate parameter is notated as μ.

Γ is the type change rate matrix, and δ is a scaling parameter for this matrix. Alternatively, a single type change rate γ can be provided for all types, in which case the model will assume uniform transitions between types.

Sampling probabilities are notated as ρ (survival sampling probability) and σ (death sampling probability).

StatsAPI.loglikelihood — Function

loglikelihood(model::AbstractBranchingProcess, tree::TreeNode, present_time; kw...)
loglikelihood(model::AbstractBranchingProcess, trees::Vector{TreeNode}, present_time; kw...)

A slight deviation from StatsAPI, as observations are not stored in the model object, so they must be passed as an argument.

Ensure that present_time > 0, since the process is defined to start at time 0. Keyword arguments reltol and abstol are passed to the ODE solver.

gcdyn.λ — Function

λ(model::AbstractBranchingProcess, type)

Evaluates the model's birth rate function at the given type.

gcdyn.μ — Function

μ(model::AbstractBranchingProcess, type)

Evaluates the model's death rate function at the given type.

gcdyn.γ — Function

γ(model::AbstractBranchingProcess, type)

Evaluates the model's type change rate function at the given type (the rate of change out of type).

γ(model::AbstractBranchingProcess, from_type, to_type)

Evaluates the model's rate of type change from from_type to to_type.