Other

Trees

gcdyn.mutate! — Function

mutate!(node, model)

Change the type of the given node according to the transition probabilities or rates in the given model.

gcdyn.sample_child! — Function

sample_child!(parent, model, present_time)

Append a new child event to the given parent node, selecting between events from EVENTS according to the given model's rate parameters.

Note that the child will a time t larger than its parent. This aligns us with the branching process models, but contradicts the notion of time typically used in phylogenetics.

gcdyn.attach! — Function

attach!(parent::TreeNode, child::TreeNode)

Adds child to parent.children and sets child.up to parent.

gcdyn.detach! — Function

detach!(parent::TreeNode, child::TreeNode)

Removes child from parent.children and sets child.up to nothing.

Base.delete! — Method

delete!(node::TreeNode)

Removes this node from its place in the tree, attaching its children to its parent.

gcdyn.map_types — Function

map_types(tree, mapping; prune_self_loops = true)

Returns a new tree like tree, but replaces the type attribute of each node with the result of the callable mapping applied to the type value.

Optionally but by default, if the map results in a type change event with the same type as its parent, the type change event is pruned from the tree.

gcdyn.map_types! — Function

map_types!(tree, mapping; prune_self_loops = true)

Replaces the type attribute of all nodes in tree with the result of the callable mapping applied to the type value.

Note that TreeNode{T} objects have a fixed typeof(node.type) == T, so the mapping must be compliant with this.

Optionally but by default, if the map results in a type change event with the same type as its parent, the type change event is pruned from the tree.

Branching processes

gcdyn.expit — Function

expit(x)

$\frac{1}{1 + exp(-x)}$

gcdyn.sigmoid — Function

sigmoid(x, xscale, xshift, yscale, yshift)

$\frac{\text{yscale}}{1 + exp(-(\text{xscale} * (x - ext{{xshift}})))} + \text{yshift}$

gcdyn.dp_logq_dt! — Function

See equations (1) and (2) of the paper.

gcdyn.dp_dt! — Function

See equation (1) of the paper.

gcdyn.type_space_index — Function

type_space_index(model::AbstractBranchingProcess, type)

Returns the index of type in the given model's type space, or nothing if not found.

Additional likelihood functions

gcdyn.naive_loglikelihood — Function

naive_loglikelihood(model, tree)

Compute the log-likelihood of the given model under the assumption that the given tree is fully observed (ie. all nodes are sampled, dead or not).

gcdyn.stadler_appx_loglikelihood — Function

stadler_appx_loglikelhood(model, tree, present_time)

Compute the log-likelihood of the given model under the assumption that all type change nodes have at least one sampled descendant.

Barido-Sottani, Joëlle, Timothy G Vaughan, and Tanja Stadler. “A Multitype Birth–Death Model for Bayesian Inference of Lineage-Specific Birth and Death Rates.” Edited by Adrian Paterson. Systematic Biology 69, no. 5 (September 1, 2020): 973–86. https://doi.org/10.1093/sysbio/syaa016.

gcdyn.stadler_appx_unconditioned_loglikelihood — Function

stadler_appx_unconditioned_loglikelhood(model, tree, present_time)

Compute the log-likelihood of the given model under the assumption that all type change nodes have at least one sampled descendant.

Does not condition on all trees having at least one sampled descendant (ie. rand_tree has keyword argument reject_stubs=False).

Barido-Sottani, Joëlle, Timothy G Vaughan, and Tanja Stadler. “A Multitype Birth–Death Model for Bayesian Inference of Lineage-Specific Birth and Death Rates.” Edited by Adrian Paterson. Systematic Biology 69, no. 5 (September 1, 2020): 973–86. https://doi.org/10.1093/sysbio/syaa016.